+++ CHANGELOG.md
@@ -1,3 +1,42 @@
+## [v2.69.0](https://github.com/docling-project/docling-core/releases/tag/v2.69.0) - 2026-03-09
+
+### Feature
+
+* Loosen dependency version constraints ([#534](https://github.com/docling-project/docling-core/issues/534)) ([`4eb0d20`](https://github.com/docling-project/docling-core/commit/4eb0d20d0455952e55f6639b0ca96a7a448453f0))
+
+## [v2.68.0](https://github.com/docling-project/docling-core/releases/tag/v2.68.0) - 2026-03-07
+
+### Feature
+
+* Add plain-text serializer ([#522](https://github.com/docling-project/docling-core/issues/522)) ([`e363c95`](https://github.com/docling-project/docling-core/commit/e363c951d8b67b67a23e7bd4a7d92969a7875a33))
+
+### Fix
+
+* Prevent infinite loop in LineBasedTokenChunker with unbreakable tokens ([#533](https://github.com/docling-project/docling-core/issues/533)) ([`a661bb1`](https://github.com/docling-project/docling-core/commit/a661bb10cb46a7b8ce338773e6e95e558b96d669))
+
+## [v2.67.1](https://github.com/docling-project/docling-core/releases/tag/v2.67.1) - 2026-03-05
+
+### Fix
+
+* Prevent hang in export_to_markdown() on nested RichTableCells ([#525](https://github.com/docling-project/docling-core/issues/525)) ([`2debe08`](https://github.com/docling-project/docling-core/commit/2debe0836fb0365e36d2b6e5b0e1c409c85f6ecb))
+
+## [v2.67.0](https://github.com/docling-project/docling-core/releases/tag/v2.67.0) - 2026-03-04
+
+### Feature
+
+* Table aware chunking ([#527](https://github.com/docling-project/docling-core/issues/527)) ([`ea359bc`](https://github.com/docling-project/docling-core/commit/ea359bcc63e0cc25d323374a234e13142adf9110))
+
+## [v2.66.0](https://github.com/docling-project/docling-core/releases/tag/v2.66.0) - 2026-02-26
+
+### Feature
+
+* Add WebVTT export and save functionality ([#523](https://github.com/docling-project/docling-core/issues/523)) ([`b8ef7ba`](https://github.com/docling-project/docling-core/commit/b8ef7bad1b7281aca0cc4b8534da62f487da6c8b))
+
+### Fix
+
+* Rich table triplet serialization ([#425](https://github.com/docling-project/docling-core/issues/425)) ([`c566268`](https://github.com/docling-project/docling-core/commit/c566268e0a3f681de0e3394446b54edaecf53b4a))
+* Support single-column table default serialization ([#526](https://github.com/docling-project/docling-core/issues/526)) ([`73b0757`](https://github.com/docling-project/docling-core/commit/73b07572cf7fbcf8b011c0280dc4b5281243350b))
+
 ## [v2.65.2](https://github.com/docling-project/docling-core/releases/tag/v2.65.2) - 2026-02-23
 
 ### Fix